backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.86.0",
    "assemblyai>=0.58.0",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "supabase>=2.28.3",
    "uvicorn>=0.42.0",
    "yt-dlp>=2026.3.17",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
